+++ CHANGELOG.md
@@ -5,13 +5,19 @@ All notable changes to the "vapor-vscode" extension will be documented in this f
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2025-04-02
+
+### Fixed
+
+- Ignore `<style>` tags when formatting Leaf files.
+
 ## [0.0.5] - 2025-04-02
 
 ### Fixed
 
 - Highlight Leaf when embedded in JavaScript code blocks.
-- Ignore script tags when formatting Leaf files.
-- Update the required version of VSCode to ^1.96.2
+- Ignore `<script>` tags when formatting Leaf files.
+- Update the required version of VS Code to 1.96.2.
 
 ## [0.0.4] - 2025-02-27